+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Add support for reading out driving licences
+- Add support for reading out ID-cards
+
+### Fix
+- Padding issues in disclosure screens
 
 ## [7.11.2] - 2025-11-21
 ### Changed